+++ src/fastoad/models/performances/mission/segments/altitude_change.py
