+++ src/fastoad/models/performances/mission/segments/speed_change.py
